Update default values for graphql-query-size and rest-api-page-size in create-asset-download-lists-by-package action

+++ .github/actions/upload-maven-package-from-json/upload-maven-package-from-json.mjs
@@ -66,7 +66,11 @@ const signatureTypes = ['md5', 'sha1', 'sha256'];
 
 const isSignatureFile = (fileName) => {
     return signatureTypes.some(type => fileName.toLowerCase().endsWith(type));
-}
+};
+
+const isSnapshot = (versionName) => {
+    return versionName.includes('SNAPSHOT');
+};
 
 const downloadFile = async (fileUrl, filePath, token) => {
     const writer = fs.createWriteStream(filePath);
@@ -229,15 +233,20 @@ const retryUpload = async (uploadUrl, fileStream, headers, maxRetries = 5, retry
             name: packageImportJson.name
         };
 
-        // Use downloadFile to download the maven-metadata.xml for the version
-        const metadataVersionFile = `maven-metadata-${version.version}.xml`;
-        const metadataVersionFilePath = `${rootDirectory}/${metadataVersionFile}`;
-        const mavenMetadataVersionUrl = `${baseUrl}/${packageImportJson.repositoryFullName}/${packageImportJson.name.replace('.', '/')}/${version.version}/maven-metadata.xml`;
-        try {
-            await downloadFile(mavenMetadataVersionUrl, metadataVersionFile, fromToken);
-        }
-        catch (error) {
-            console.error(`Error downloading version maven-metadata.xml file from ${mavenMetadataVersionUrl}.`);
+        let metadataVersionFile, metadataVersionFilePath, mavenMetadataVersionUrl;
+
+        if (isSnapshot(version.version)) {
+            // Use downloadFile to download the maven-metadata.xml for the version
+            metadataVersionFile = `maven-metadata-${version.version}.xml`;
+            metadataVersionFilePath = `${rootDirectory}/${metadataVersionFile}`;
+            mavenMetadataVersionUrl = `${baseUrl}/${packageImportJson.repositoryFullName}/${packageImportJson.name.replaceAll('.', '/')}/${version.version}/maven-metadata.xml`;
+            try {
+                await downloadFile(mavenMetadataVersionUrl, metadataVersionFile, fromToken);
+            }
+            catch (error) {
+                console.error(`Error downloading version maven-metadata.xml file from ${mavenMetadataVersionUrl}.`);
+                console.error(error);
+            }
         }
 
         // Get the files for the package version
@@ -253,7 +262,7 @@ const retryUpload = async (uploadUrl, fileStream, headers, maxRetries = 5, retry
             let fileResponse;
             let file = files[j];
             let fileName = file.name;
-            let uploadUrl = `${baseUrl}/${packageImportJson.toOwner}/${packageImportJson.repository}/${packageImportJson.name.replace('.', '/')}/${version.version}/${fileName}`;
+            let uploadUrl = `${baseUrl}/${packageImportJson.toOwner}/${packageImportJson.repository}/${packageImportJson.name.replaceAll('.', '/')}/${version.version}/${fileName}`;
             let filePath = `${rootDirectory}/${fileName}`;
 
             try {
@@ -265,7 +274,7 @@ const retryUpload = async (uploadUrl, fileStream, headers, maxRetries = 5, retry
                 console.log(`\tTokens refreshed. Version ${version.version} has ${files.length} files.`);
                 file = files[j];
                 fileName = file.name;
-                uploadUrl = `${baseUrl}/${packageImportJson.toOwner}/${packageImportJson.repository}/${packageImportJson.name.replace('.', '/')}/${version.version}/${fileName}`;
+                uploadUrl = `${baseUrl}/${packageImportJson.toOwner}/${packageImportJson.repository}/${packageImportJson.name.replaceAll('.', '/')}/${version.version}/${fileName}`;
                 filePath = `${rootDirectory}/${fileName}`;
                 fileResponse = await downloadFile(file.url, filePath, fromToken);
             }
@@ -329,39 +338,49 @@ const retryUpload = async (uploadUrl, fileStream, headers, maxRetries = 5, retry
             });
         }
 
-        // Remove the signatureTypes from the version metadata file and upload to the destination repository
-        try {
-            removeSignatureTypesFromMavenVersionMetadata(metadataVersionFilePath);
+        if (isSnapshot(version.version)) {
+            // Remove the signatureTypes from the version metadata file and upload to the destination repository
+            try {
+                removeSignatureTypesFromMavenVersionMetadata(metadataVersionFilePath);
 
-            console.log(`\tMaven version metadata for version ${version.version} updated.`);
+                console.log(`\tMaven version metadata for version ${version.version} updated.`);
 
-            const metadataVersionUploadUrl = `${baseUrl}/${packageImportJson.toOwner}/${packageImportJson.repository}/${packageImportJson.name.replace('.', '/')}/${version.version}/${metadataVersionFile}`;
-            const metadataVersionFileStream = fs.createReadStream(metadataVersionFilePath);
+                const metadataVersionUploadUrl = `${baseUrl}/${packageImportJson.toOwner}/${packageImportJson.repository}/${packageImportJson.name.replaceAll('.', '/')}/${version.version}/maven-metadata.xml`;
+                const metadataVersionFileStream = fs.createReadStream(metadataVersionFilePath);
 
-            const uploadResponse = await retryUpload(metadataVersionUploadUrl, metadataVersionFileStream, {
-                headers: {
-                    Authorization: `Bearer ${toToken}`,
-                    'Content-Type': 'application/octet-stream',
-                    'Content-Length': fs.statSync(metadataVersionFilePath).size
-                }
-            });
+                const uploadResponse = await retryUpload(metadataVersionUploadUrl, metadataVersionFileStream, {
+                    headers: {
+                        Authorization: `Bearer ${toToken}`,
+                        'Content-Type': 'application/octet-stream',
+                        'Content-Length': fs.statSync(metadataVersionFilePath).size
+                    }
+                });
 
-            console.log(`\tVersion metadata for ${version.version} uploaded.`);
+                console.log(`\tVersion metadata for ${version.version} uploaded.`);
 
-            //Delete the file
-            fs.unlink(metadataVersionFilePath, (err) => {
-                if (err) {
-                    console.error(`\t\tFailed to delete local copy of ${metadataVersionFilePath}`);
-                    console.error(err);
-                }
-            });
+                //Delete the file
+                fs.unlink(metadataVersionFilePath, (err) => {
+                    if (err) {
+                        console.error(`\t\tFailed to delete local copy of ${metadataVersionFilePath}`);
+                        console.error(err);
+                    }
+                });
+            }
+            catch (error) {
+                console.error(`\t${metadataVersionFile} failed to upload.`);
+                console.error(error);
+            }
 
-            console.log(`Version ${version.version} migration of files and metadata complete.`);
         }
-        catch (error) {
-            console.error(`\t${metadataVersionFile} failed to upload.`);
-            console.error(error);
-        }
+
+        fs.unlink(`${rootDirectory}/maven-metadata.xml`, (err) => {
+            if (err) {
+                console.error(`\tFailed to delete local copy of maven-metadata.xml`);
+                console.error(err);
+            }
+        });
+        
+        console.log(`Version ${version.version} migration complete.`);
     }
 
     core.setOutput('results', results);
